add new contract: personal.bos ,which contains actions:
1.setpersonal
2.sethomepage
this contract has a table "personaldata" ,which's a key-value map to store customized public info for accounts. it has 3 columns:
   key: the primary uint64_t key, which's compressed from "readablekey"
   readablekey: user input "key",should be shorter than 12, will use it to generate "key"
   value: the value, should be less than 1024.
setpersonal:user can call to set the key-value they wish,and read it via get table.
sethomepage:the same with setpersonal, just the key is fixed to "homepage", by supplying homepage, dapp can easily guide user to their homepage.
            with commits on "boscore/bos",user can also use "cleos set personal ..." and "cleos get account ..." to set personaldata and read homepage.
also add an inline interface "getpersonal" in personal.bos.hpp for easily accessing this multi-index for contracts.

Deployment:
when deploying,you need to create account "personal.bos" and set this contract to it to let it working.